added

+++ graphs_contd.cpp
@@ -29,7 +29,7 @@ public:
 		int *dist = new int[V];
 		//Init all nodes with inf dist
 		for(int i=0;i<V;i++){
-			dist[i] = INT_MAX;
+			dist[i] = INT8_MAX;
 		}
 		dist[s] = 0;
 
@@ -38,7 +38,7 @@ public:
 			q.pop();
 
 			for(auto child:l[node]){
-				if(dist[child]==INT_MAX){
+				if(dist[child]==INT8_MAX){
 					q.push(child);
 					dist[child] = dist[node] + 1;
 				}
